# backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.ChangePasswordRequest:
    properties:
      confirm_new_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
  controllers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.ResetPasswordRequest:
    properties:
      confirm_new_password:
        type: string
      new_password:
        type: string
    type: object
  controllers.SetupCredentialRequest:
    properties:
      confirm_new_password:
        type: string
      new_email:
        type: string
      new_name:
        type: string
      new_password:
        type: string
    type: object
  controllers.VerifyOTPRequest:
    properties:
      otp_code:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
  description: API Backend untuk aplikasi SIUJI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIUJI API Documentation
  version: "1.0"
paths:
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change current user password by providing old and new password
      parameters:
      - description: Old and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Change failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Change password
      tags:
      - Auth
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send an OTP verification code to the registered email
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset OTP
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Clear authentication cookies and log out user
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
      summary: User logout
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Retrieve logged-in user profile details
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      summary: Get current user profile
      tags:
      - Auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using new password after OTP verification
      parameters:
      - description: New password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Reset failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Reset password
      tags:
      - Auth
  /api/v1/auth/setup-credential:
    post:
      consumes:
      - application/json
      description: Setup new name, email, and password for first-time login users
      parameters:
      - description: New credentials data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SetupCredentialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Credential setup successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request / setup failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Setup initial credentials
      tags:
      - Auth
  /api/v1/auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to email for password reset
      parameters:
      - description: OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Verification failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Verify OTP code
      tags:
      - Auth
swagger: "2.0"